# shared/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Hockey Club Platform API
  version: 0.1.0
  description: |
    Phase 1 (MVP) API: accounts & roles, player profiles, availability,
    lineups and the suggestion engine. Schemas mirror shared/contracts/types.ts.
    FROZEN — change deliberately and keep types.ts in sync.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/magic-link:
    post:
      summary: Send a magic sign-in link by email or SMS
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                phone: { type: string }
      responses:
        '202':
          description: |
            Always 202, whether or not the email is registered, to avoid account enumeration.
            In development only, the response includes `devLink` so you can sign in without email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
                  devLink: { type: string }
  /auth/magic-link/verify:
    post:
      summary: Exchange a magic-link token for a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: 'Signed in. Send "Authorization: Bearer <accessToken>" on later requests.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '401': { description: Invalid, expired or already-used token }
...
  /members/{memberId}/invite:
    parameters: [{ $ref: '#/components/parameters/memberId' }]
    post:
      summary: Email a member a magic link to join (manager of one of their teams)
      responses:
        '202': { description: Invite sent }
        '403': { description: Not a manager of this member }
...
  /membership-plans/{planId}/checkout:
    parameters: [{ name: planId, in: path, required: true, schema: { type: string } }]
    post:
      summary: Start a Stripe Checkout for a membership plan; returns the page to send the member to
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutSession' }
        '503': { description: Payments are not set up yet }
  /webhooks/stripe:
    post:
      summary: Stripe webhook (signature-verified; records payments and updates memberships)
      responses:
        '200': { description: Event handled or ignored }
        '400': { description: Bad signature }
        '503': { description: Payments are not set up yet }
  /clubs/{clubId}/membership-plans:
    parameters: [{ $ref: '#/components/parameters/clubId' }]
    post:
      summary: Add a membership plan (club admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewMembershipPlan' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MembershipPlan' }
    get:
      summary: Membership plans offered by the club
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MembershipPlan' } }
  /clubs/{clubId}/teams:
    parameters: [{ $ref: '#/components/parameters/clubId' }]
    get:
      summary: List teams in a club
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Team' } }
  /teams/{teamId}/players:
    parameters: [{ $ref: '#/components/parameters/teamId' }]
    get:
      summary: Squad for a team
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/PlayerProfile' } }
    post:
      summary: Add a new player to the squad (manager/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewPlayer' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlayerProfile' }
  /teams/{teamId}/players/{memberId}:
    parameters:
      - { $ref: '#/components/parameters/teamId' }
      - { $ref: '#/components/parameters/memberId' }
    patch:
      summary: Edit positions / ratings (manager/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlayerProfileUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlayerProfile' }
  /teams/{teamId}/fixtures:
    parameters: [{ $ref: '#/components/parameters/teamId' }]
    get:
      summary: Fixtures for a team, soonest first
      parameters:
        - name: from
          in: query
          description: Only fixtures starting at or after this time
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Fixture' } }
  /fixtures/{fixtureId}/availability:
    parameters: [{ $ref: '#/components/parameters/fixtureId' }]
    get:
      summary: Availability of every squad member for a fixture
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Availability' } }
    put:
      summary: Set my (or my child's) availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, status]
              properties:
                memberId: { type: string }
                status: { $ref: '#/components/schemas/AvailabilityStatus' }
                note: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Availability' }
  /formations:
    get:
      summary: Built-in formations, optionally filtered by squad format
      parameters:
        - name: format
          in: query
          schema: { $ref: '#/components/schemas/SquadFormat' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Formation' } }
  /teams/{teamId}/formations:
    parameters: [{ $ref: '#/components/parameters/teamId' }]
    get:
      summary: This team's custom formations, optionally filtered by squad format
      parameters:
        - name: format
          in: query
          schema: { $ref: '#/components/schemas/SquadFormat' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Formation' } }
    post:
      summary: Save a custom formation from typed line counts (manager/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewCustomFormation' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Formation' }
        '400':
          description: Line counts don't total 5, 7, or 11 including the keeper
  /teams/{teamId}/formation-layouts:
    parameters: [{ $ref: '#/components/parameters/teamId' }]
    get:
      summary: This team's saved position adjustments, one per formation they've adjusted
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/FormationLayout' } }
  /teams/{teamId}/formation-layouts/{formationId}:
    parameters:
      - { $ref: '#/components/parameters/teamId' }
      - { name: formationId, in: path, required: true, schema: { type: string } }
    put:
      summary: Save adjusted positions for a formation (manager/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [positions]
              properties:
                positions:
                  type: object
                  additionalProperties: { $ref: '#/components/schemas/PitchPosition' }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormationLayout' }
        '400':
          description: Unknown formation, or a position id that isn't in it
    delete:
      summary: Reset a formation back to its automatic layout (manager/admin)
      responses:
        '204': { description: Reset }
  /fixtures/{fixtureId}/lineup:
    parameters: [{ $ref: '#/components/parameters/fixtureId' }]
    get:
      summary: Current lineup for a fixture
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lineup' }
        '404': { description: No lineup yet }
    put:
      summary: Save the lineup (manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Lineup' }
      responses:
        '200': { description: Saved }
  /fixtures/{fixtureId}/lineup/suggest:
    parameters: [{ $ref: '#/components/parameters/fixtureId' }]
    post:
      summary: Ask the suggestion engine for a lineup + substitution plan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SuggestionRequest' }
      responses:
        '200':
          description: Suggestion (not saved until the manager saves it)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuggestionResult' }
  /fixtures/{fixtureId}/lineup/share:
    parameters: [{ $ref: '#/components/parameters/fixtureId' }]
    post:
      summary: Share the saved lineup with selected players (default everyone in it)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                memberIds:
                  type: array
                  items: { type: string }
                  description: Omit to share with the whole squad
      responses:
        '202':
          description: Notifications sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sharedWith: { type: integer }
        '409': { description: No saved lineup to share }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    clubId: { name: clubId, in: path, required: true, schema: { type: string } }
    teamId: { name: teamId, in: path, required: true, schema: { type: string } }
    memberId: { name: memberId, in: path, required: true, schema: { type: string } }
    fixtureId: { name: fixtureId, in: path, required: true, schema: { type: string } }
  schemas:
...
    SquadFormat: { type: integer, enum: [5, 7, 11] }
...
    AvailabilityStatus: { type: string, enum: [available, unavailable, maybe, no_response] }
...
    Team:
      type: object
      required: [id, clubId, name, ageGroup, defaultFormat]
      properties:
        id: { type: string }
        clubId: { type: string }
        name: { type: string }
        ageGroup: { $ref: '#/components/schemas/AgeGroup' }
        defaultFormat: { $ref: '#/components/schemas/SquadFormat' }
    PlayerProfile:
      type: object
      required: [memberId, displayName, positions, skill, stamina, seasonMinutes]
      properties:
        memberId: { type: string }
        displayName: { type: string }
        shirtNumber: { type: integer }
        positions: { type: array, items: { $ref: '#/components/schemas/PositionLine' } }
        skill: { type: integer, minimum: 1, maximum: 10 }
        stamina: { type: integer, minimum: 1, maximum: 10 }
        seasonMinutes: { type: integer, minimum: 0 }
    Fixture:
      type: object
      required: [id, teamId, opponent, startsAt, venue, homeAway, format, durationMinutes, periods]
      properties:
        id: { type: string }
        teamId: { type: string }
        opponent: { type: string }
        startsAt: { type: string, format: date-time }
        venue: { type: string }
        homeAway: { type: string, enum: [home, away] }
        format: { $ref: '#/components/schemas/SquadFormat' }
        durationMinutes: { type: integer }
        periods: { type: integer }
    Availability:
      type: object
      required: [fixtureId, memberId, status, updatedAt]
      properties:
        fixtureId: { type: string }
        memberId: { type: string }
        status: { $ref: '#/components/schemas/AvailabilityStatus' }
        note: { type: string }
        updatedAt: { type: string, format: date-time }
...
    Formation:
      type: object
      required: [id, name, format, slots]
      properties:
        id: { type: string }
        name: { type: string }
        format: { $ref: '#/components/schemas/SquadFormat' }
        slots: { type: array, items: { $ref: '#/components/schemas/FormationSlot' } }
        teamId: { type: string, description: "Present only for a team's saved custom formation" }
...
    FormationLayout:
      type: object
      required: [formationId, positions]
      properties:
        formationId: { type: string }
        positions:
          type: object
          description: Per position id; overrides the automatic layout for that position
          additionalProperties: { $ref: '#/components/schemas/PitchPosition' }
...
    NewPlayer:
      type: object
      required: [firstName, lastName, positions, skill, stamina]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        displayName: { type: string, description: 'Defaults to "First L."' }
        shirtNumber: { type: integer }
        positions: { type: array, items: { $ref: '#/components/schemas/PositionLine' } }
        skill: { type: integer, minimum: 1, maximum: 10 }
        stamina: { type: integer, minimum: 1, maximum: 10 }
    PlayerProfileUpdate:
      type: object
      properties:
        displayName: { type: string }
        shirtNumber: { type: integer }
        positions: { type: array, items: { $ref: '#/components/schemas/PositionLine' } }
        skill: { type: integer, minimum: 1, maximum: 10 }
        stamina: { type: integer, minimum: 1, maximum: 10 }
        seasonMinutes: { type: integer, minimum: 0 }
...
    AuthSession:
      type: object
      required: [accessToken, me]
      properties:
        accessToken: { type: string }
        me: { $ref: '#/components/schemas/Me' }
    MembershipPlan:
      type: object
      required: [id, clubId, name, amountPence, interval]
      properties:
        id: { type: string }
        clubId: { type: string }
        name: { type: string }
        amountPence: { type: integer }
        interval: { type: string, enum: [month, quarter, year] }
    NewMembershipPlan:
      type: object
      required: [name, amountPence, interval]
      properties:
        name: { type: string }
        amountPence: { type: integer, minimum: 1 }
        interval: { type: string, enum: [month, quarter, year] }
...
    CheckoutSession:
      type: object
      properties:
        url: { type: string }
        demo: { type: boolean }
...